+++ examples/sdmmc.rs
@@ -11,6 +11,7 @@
 mod utilities;
 
 use cortex_m_rt::entry;
+use embedded_hal_1::delay::DelayNs; // this example uses embedded-hal v1.0
 use stm32h7xx_hal::gpio::Speed;
 use stm32h7xx_hal::sdmmc::{SdCard, Sdmmc};
 use stm32h7xx_hal::{pac, prelude::*};
@@ -114,7 +115,7 @@ fn main() -> ! {
 
         info!("Waiting for card...");
 
-        delay.delay_ms(1000u32);
+        delay.delay_ms(1000);
         led.toggle();
     }