+++ examples/serial-halfduplex.rs
@@ -71,15 +71,15 @@ fn main() -> ! {
     loop {
         // Tx from usart3
         for c in "Hello, world!".chars() {
-            block!(usart3.write(c as u8)).unwrap();
-            let received = block!(usart6.read()).unwrap();
+            block!(usart3.write_byte(c as u8)).unwrap();
+            let received = block!(usart6.read_byte()).unwrap();
             info!("usart6 rx {}", received as char);
         }
 
         // Tx from usart6
         for c in "Hello, world!".chars() {
-            block!(usart6.write(c as u8)).unwrap();
-            let received = block!(usart3.read()).unwrap();
+            block!(usart6.write_byte(c as u8)).unwrap();
+            let received = block!(usart3.read_byte()).unwrap();
             info!("usart3 rx {}", received as char);
         }
     }