+++ src/gpio.rs
@@ -76,8 +76,9 @@ pub use exti::ExtiPin;
 mod dynamic;
 pub use dynamic::{Dynamic, DynamicPin};
 mod hal_02;
+mod hal_1;
 
-pub use embedded_hal::digital::v2::PinState;
+pub use embedded_hal_1::digital::PinState;
 
 use core::fmt;
 
@@ -458,7 +459,6 @@ impl<const P: char, const N: u8, MODE> Pin<P, N, Output<MODE>> {
     pub fn set_low(&mut self) {
         self._set_low()
     }
-
     /// Is the pin in drive high or low mode?
     #[inline(always)]
     pub fn get_state(&self) -> PinState {