+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 ## [Unreleased]
 
 * MSRV increased to Rust 1.66.1 [#473]
+* Upgraded to embedded-hal v1.0.0-rc.3
+
+* delay: `delay_ms` and 'delay_us` methods now require importing: `use embedded_hal::delay::DelayNs;`
+
+* pwm: Renamed `get_max_duty` -> `max_duty_cycle`; `set_duty` -> `set_duty_cycle`
+* pwm: `enable` method now returns type `Result<(), PwmError>`
+* pwm: LPTIMs can return a `NotEnabled` error from `set_duty_cycle`
+
+* serial: Implement embedded-io `Read` and `Write` traits
+* serial: Rename methods `read` -> `read_byte`; `write` -> `write_byte`
+
+* rng: Use the `fill` method instead of the old `read` method
 
 ## [v0.15.1] 2023-11-03